Created a new module called limits.
