+++ src/pyspextool/plot/__init__.py
@@ -2,4 +2,4 @@
 This subpackage contains functions for plot.
 """
 
-__all__ = ['buffer_range', 'get_image_range', 'get_spec_range']
+__all__ = ['limits']
